+++ src/index.tsx
@@ -1,14 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+
 import Router from "./Router";
 
-import "./index.scss";
-
-const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
-);
-root.render(
-  // <React.StrictMode>
-  <Router />
-  // </React.StrictMode> */}
+ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
+  <React.StrictMode>
+    <Router />
+  </React.StrictMode>
 );